Moved solutions to P1 file, Math now holds single fcn - TESTS FAIL
Had originally expected to reuse the sumOfMultiplesLessThan function and thus it
was placed in the 'math' file.  However, I was also experimenting with various
solutions and that started to clog the 'math' file with suboptimal
implementations.

Pulled all suboptimal solutions out to the P1 file.  Added the constant time
solution helper function to 'math'