Collapsed P1 test cases into forEach around all methods tried
I have been trying different solutions to the problem and was copying in the
same test cases.  This was tedious and messy, so I extracted out the tests and
expected values so that I could write the setup once.

Now adding a new method is as easy as adding it to the array of 'solutions'

Similarly, if we want to try another expected result, it can be added to the
'answers' array.

By adding a solution or answer, ALL methods found in 'solutions' will be tested